pyproject.toml
[tool.poetry]
name = "crawling-api"
version = "0.1.0"
description = ""
authors = ["k.iimori <keita.iimori@gmail.com>"]

[tool.taskipy.tasks]
dev = "uvicorn src.main:app --host 0.0.0.0 --port 8000 --access-log --reload"
lint = "bash -c 'isort ./src && black ./src && flake8 ./src'"
fix = "bash -c 'isort ./src ; black ./src'"
export = "poetry export --without-hashes --format=requirements.txt --output=./src/requirements.txt"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.115.2"
uvicorn = "^0.32.0"
mangum = "^0.19.0"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
pyyaml = "^6.0.2"
beautifulsoup4 = "^4.12.3"
html2text = "^2024.2.26"

[tool.poetry.group.dev.dependencies]
urllib3 = "1.26.18"
mypy = "^1.12.1"
pytest = "^5.2"
poetry-plugin-export = "^1.8.0"
isort = "^5.13.2"
black = "^24.10.0"
flake8 = "^7.1.1"
taskipy = "^1.13.0"
pydantic = {extras = ["email"], version = "^2.9.2"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
